requirements.txt
django-widget-tweaks 